refactor(main): extract screenshot detector out of index.ts
