+++ src/main/screenshotDetector.ts
@@ -1,0 +1,104 @@
+// ===== Screenshot detector =====
+//
+// Two parallel detection paths — together they cover all four macOS screenshot
+// shortcuts without any native module:
+//
+// 1. Clipboard poller (500ms): catches ⌘⇧Control+3/4 (clipboard-only shots).
+//    Tracks clipboard "generation" by hashing availableFormats() + image size;
+//    when a new image appears it pushes screenshotDetected to the renderer.
+//    We intentionally do NOT read the full pixel buffer every tick — only
+//    formats + size, so the poll is cheap. The renderer calls uploadBuffer
+//    which reads the clipboard once via a separate IPC.
+//
+// 2. Desktop folder watcher: catches ⌘⇧3/4 (file-only shots). macOS names
+//    them "Screenshot YYYY-MM-DD at HH.MM.SS.png". The fs.watch callback
+//    fires on the rename event (file creation) and we filter by that pattern.
+//    We push the absolute path so the renderer can scp it directly via
+//    uploadFiles (no extra read needed).
+//
+// Both paths are no-ops when no mainWindow exists. The Desktop watcher is
+// started once at app-ready and never restarted (the Desktop path doesn't
+// change). The clipboard poller is the same.
+
+import { clipboard } from "electron";
+import { basename, join } from "node:path";
+import { watch as fsWatch } from "node:fs";
+import { homedir } from "node:os";
+import { IPC } from "../shared/types.js";
+
+const SCREENSHOT_RE = /^Screenshot \d{4}-\d{2}-\d{2} at \d{2}\.\d{2}\.\d{2}.*\.png$/i;
+
+let screenshotClipboardTimer: ReturnType<typeof setInterval> | null = null;
+let screenshotDesktopWatcher: ReturnType<typeof fsWatch> | null = null;
+
+// Cheap fingerprint: join of available formats + "<w>x<h>". We don't hash
+// pixel data — just enough to detect "something new appeared".
+function clipboardImageFingerprint(): string {
+  const fmts = clipboard.availableFormats().join(",");
+  if (!fmts.includes("image")) return "";
+  const img = clipboard.readImage();
+  if (img.isEmpty()) return "";
+  const { width, height } = img.getSize();
+  return `${fmts}|${width}x${height}`;
+}
+
+export function startScreenshotDetector(
+  send: (channel: string, payload: unknown) => void,
+): void {
+  // macOS-only: the clipboard fingerprint relies on Mac's screencapture
+  // populating the clipboard on ⌘⇧^3/4, and the file watcher targets the
+  // Mac-default `~/Desktop/Screenshot YYYY-MM-DD at HH.MM.SS.png` filename.
+  // On Linux/Windows both paths would no-op at best and burn CPU on the
+  // 500ms clipboard poller at worst.
+  if (process.platform !== "darwin") return;
+
+  // --- Clipboard poller ---
+  let lastFingerprint = clipboardImageFingerprint();
+  screenshotClipboardTimer = setInterval(() => {
+    const fp = clipboardImageFingerprint();
+    if (fp && fp !== lastFingerprint) {
+      lastFingerprint = fp;
+      send(IPC.screenshotDetected, { source: "clipboard" });
+    }
+  }, 500);
+
+  // --- Desktop folder watcher ---
+  const desktop = join(homedir(), "Desktop");
+  try {
+    screenshotDesktopWatcher = fsWatch(desktop, (event, filename) => {
+      if (event !== "rename" || !filename) return;
+      if (!SCREENSHOT_RE.test(basename(filename))) return;
+      const fullPath = join(desktop, filename);
+      // Small delay — fs.watch fires on the rename (inode creation) before
+      // the file is fully written. 300ms is enough for a PNG flush.
+      setTimeout(() => {
+        send(IPC.screenshotDetected, {
+          source: "file",
+          path: fullPath,
+        });
+      }, 300);
+    });
+  } catch {
+    // Desktop might not exist or be accessible (e.g. on Linux in dev).
+    // Silently skip — clipboard poller still works.
+  }
+}
+
+export function stopScreenshotDetector(): void {
+  if (screenshotClipboardTimer) {
+    clearInterval(screenshotClipboardTimer);
+    screenshotClipboardTimer = null;
+  }
+  if (screenshotDesktopWatcher) {
+    screenshotDesktopWatcher.close();
+    screenshotDesktopWatcher = null;
+  }
+}
+
+// IPC: renderer calls this to read the current clipboard image as PNG bytes.
+// Separate from the poller so we only read full pixel data on demand.
+export function readClipboardImageBuffer(): Buffer | null {
+  const img = clipboard.readImage();
+  if (img.isEmpty()) return null;
+  return img.toPNG();
+}
